Add ONNX importer tests for Conv layers and Vulkan kernels
- Implement tests for Conv2d with asymmetric and symmetric padding, Conv3d, and ConvTranspose2d in `test_onnx_import.cpp` to ensure correct handling of various padding scenarios and dimensionality.
- Introduce Vulkan compute shaders for Batch Normalization backward pass, expanding operations for uint8_t types, and Welford variance calculations for different data types (Float16, Float32, Float64, BFloat16).
- Add a new `where` kernel for conditional selection based on a boolean condition.
- Document multi-dtype and multi-backend coverage gaps in `MULTIDTYPE_COVERAGE_GAPS.md` to track missing tests for critical operations across backends and data types.

+++ src/backends/cuda/kernels/conv3d.cu
@@ -875,7 +875,13 @@ auto conv3d_backward_kernel(
     } while (0)
 
     switch (input.dtype()) {
-        case DType::Float32:  TENZOR_CONV3D_BWD_DISPATCH(float,           float); break;
+        // Upgrade f32 accumulator to double. Each conv3d_backward_weight
+        // thread sums thousands of grad_output*input pairs; a float
+        // accumulator drifts by O(sqrt(N)) ULPs, which pushes the
+        // cross-backend parity test past its rtol=1e-3 threshold. Using
+        // double here costs some perf per MAC but brings CUDA in line
+        // with the CPU reference.
+        case DType::Float32:  TENZOR_CONV3D_BWD_DISPATCH(float,           double); break;
         case DType::Float64:  TENZOR_CONV3D_BWD_DISPATCH(double,          double); break;
         case DType::Float16:  TENZOR_CONV3D_BWD_DISPATCH(__half,          float); break;
         case DType::BFloat16: TENZOR_CONV3D_BWD_DISPATCH(__nv_bfloat16,   float); break;
